Move ValueInput to Common
Makes more sense for a shared component
